THE FLOATS ARE GONE.

+++ engine/src/test/battlecode/world/TestGame.java
@@ -57,7 +57,7 @@ public class TestGame {
      *
      * @return the x coordinate of the map origin.
      */
-    public float getOriginX() {
+    public int getOriginX() {
         return world.getGameMap().getOrigin().x;
     }
 
@@ -66,7 +66,7 @@ public class TestGame {
      *
      * @return the y coordinate of the map origin.
      */
-    public float getOriginY() {
+    public int getOriginY() {
         return world.getGameMap().getOrigin().y;
     }
